% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_multinom_probs.R
\name{generate_multinom_probs}
\alias{generate_multinom_probs}
\title{Generate Probabiltiies for Multinomial Draws}
\usage{
generate_multinom_probs(V = NULL, B = NULL, X = NULL, X.incl.X0 = FALSE)
}
\arguments{
\item{V}{A numeric value stating the number of categories desired.}

\item{B}{A list, each element of which contains a parameter vector. The list should
have length \code{V - 1}.}

\item{X}{A matrix, each row of which contains subject covariate/predictor values.}

\item{X.incl.X0}{Logical. When \code{TRUE}, \code{X} contains a column of 1's for
the intercept. Otherwise, a column of 1's is generated internally. Default is \code{FALSE}.}
}
\value{
A matrix containing subject-specific probabilities for each category of the
multinomial distribution. The number of rows equals \code{nrow(X)} and the number of 
columns equals \code{V}.
}
\description{
Obtain probabilities for each category of a multinomial distribution
based on covariate and parameter values based on the logit models
for the multinomial distribution.
}
\examples{
## number of categories
vt <- 3

## covariate values
xt <- matrix(rnorm(10 * 2), ncol = 2, nrow = 10)

## list of parameter vectors
bt <- list(b1 = c(1, 0.25, -0.25),
           b2 = c(-0.5, 0.15, 0.15))

## subject specific probabilities for each category
generate_multinom_probs(V = vt, X = xt, B = bt)

}
